NXDRIVE-1683: Move test_invalid_credentials() from old_functional to functional
